backend/requirements.txt
flask
Werkzeug
flask-cors
python-dotenv
certifi
pymysql